feat: add server persistence and networks endpoint; scaffold group scaler
- extend `servers` schema with `group_id`, `type`, `location` and add index on `group_id`
- add `model.Server` with `Save()` to persist server metadata
- add `GET /networks` endpoint using Hetzner API and register route in `main.go`
- align `Group` and `Template` ID fields to `int` for DB consistency
- add `services/groupScaler.go` with `ScaleUp`/`ScaleOut` placeholders

+++ model/group.go
@@ -6,8 +6,8 @@ import (
 )
 
 type Group struct {
-	Id          int64    `json:"id"`
-	TemplateId  int64    `json:"templateId"  binding:"required"`
+	Id          int      `json:"id"`
+	TemplateId  int      `json:"templateId"  binding:"required"`
 	Zone        string   `json:"zone"        binding:"required"`
 	Locations   []string `json:"locations"   binding:"required"`
 	ServerTypes []string `json:"serverTypes" binding:"required"`
